requirements.txt
appnope==0.1.0
autopep8==1.5
backcall==0.1.0
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.1
Django==2.2.10
django-cte==1.1.5
django-extensions==2.2.8
django-silk==4.0.0
-e git+https://github.com/vbudovski/django-tree.git@8f34ec3795930974170ec23abdd82532b8e83f70#egg=django_tree
django-webpack-loader==0.7.0
djangorestframework==3.11.0
gprof2dot==2019.11.30
idna==2.9
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.16.0
Jinja2==2.11.1
MarkupSafe==1.1.1
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.3
psycopg2==2.8.4
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.5.2
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
